[eth_sync] Refactor BaselineScheduler for the v2 executor (part 7/9: scheduler core)
Merges BaselineScheduler-tpp.h into BaselineScheduler.h and reworks the
scheduler core for the v2 (Ethereum executor) path:
- BaselineScheduler-tpp.h content folded into the class template (single
  definition, removed the header-only include split)
- SchedulerSerialImpl / SchedulerParallelImpl templates extended for the
  v2 scheduler entry points used by the eth_sync verifier
- baseline scheduling, batch commit, MPT state-root wiring retained for the
  v1 path

~581 valid insertions.
